+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 ##### Enhancements
 
-- Support for analyzing Swift Package Manager projects
+- Support for magic code comments to ignore unused declarations.
+- Support for analyzing Swift Package Manager projects.
 
 ##### Bug Fixes